@@ -37,27 +39,29 @@
   "license": "MIT",
   "peerDependencies": {
     "react": "^16.0",
-    "react-native": ">=0.57 <=0.59"
+    "react-native": ">=0.57 <=0.62"
   },
   "devDependencies": {
-    "@babel/core": "^7.4.3",
-    "@react-native-community/eslint-config": "^0.0.5",
+    "@babel/core": "^7.9.6",
+    "@babel/runtime": "^7.9.6",
+    "@react-native-community/eslint-config": "^1.1.0",
     "@semantic-release/git": "7.0.8",
+    "@types/react-native": "^0.62.10",
     "babel-core": "^7.0.0-bridge.0",
-    "babel-jest": "^24.7.0",
+    "babel-jest": "^26.0.1",
     "babel-plugin-module-resolver": "^3.2.0",
-    "detox": "^12.1.2",
-    "eslint": "^5.16.0",
+    "detox": "^16.5.0",
+    "eslint": "^7.0.0",
     "eslint-plugin-prettier": "^3.0.1",
-    "flow-bin": "^0.86.0",
+    "flow-bin": "^0.113.0",
     "husky": "^2.2.0",
-    "jest": "24.7.0",
-    "metro-react-native-babel-preset": "0.51.1",
+    "jest": "^26.0.1",
+    "metro-react-native-babel-preset": "^0.59.0",
     "prettier": "^1.17.0",
     "pretty-quick": "^1.10.0",
-    "react": "16.6.3",
-    "react-native": "0.58.4",
-    "react-test-renderer": "16.6.3",
+    "react": "16.11.0",
+    "react-native": "0.62.2",
+    "react-test-renderer": "16.11.0",
     "rimraf": "^2.6.3",
     "semantic-release": "15.13.3",
     "typescript": "^3.4.1"
