@@ -5,7 +5,6 @@
  * LICENSE file in the root directory of this source tree.
  *
  * @format
- * @flow
  */
 
 'use strict';
@@ -25,76 +24,27 @@ const {
 
 import CameraRoll from '../../js/CameraRoll';
 
-const groupByEveryN = require('groupByEveryN');
-const logError = require('logError');
-
-import type {
-  PhotoIdentifier,
-  PhotoIdentifiersPage,
-  GetPhotosParams,
-} from '../../js/CameraRoll';
-
-type Props = $ReadOnly<{|
-  /**
-   * The group where the photos will be fetched from. Possible
-   * values are 'Album', 'All', 'Event', 'Faces', 'Library', 'PhotoStream'
-   * and SavedPhotos.
-   */
-  groupTypes:
-    | 'Album'
-    | 'All'
-    | 'Event'
-    | 'Faces'
-    | 'Library'
-    | 'PhotoStream'
-    | 'SavedPhotos',
-
-  /**
-   * Number of images that will be fetched in one page.
-   */
-  batchSize: number,
-
-  /**
-   * A function that takes a single image as a parameter and renders it.
-   */
-  renderImage: PhotoIdentifier => React.Node,
-
-  /**
-   * imagesPerRow: Number of images to be shown in each row.
-   */
-  imagesPerRow: number,
-
-  /**
-   * A boolean that indicates if we should render large or small images.
-   */
-  bigImages?: boolean,
-
-  /**
-   * The asset type, one of 'Photos', 'Videos' or 'All'
-   */
-  assetType: 'Photos' | 'Videos' | 'All',
-|}>;
-
-type State = {|
-  assets: Array<PhotoIdentifier>,
-  data: Array<Array<?PhotoIdentifier>>,
-  seen: Set<string>,
-  lastCursor: ?string,
-  noMore: boolean,
-  loadingMore: boolean,
-|};
-
-type Row = {
-  item: Array<?PhotoIdentifier>,
+const groupByEveryN = function groupByEveryN(num) {
+  const n = num;
+  return arrayArg => {
+    const array = [...arrayArg];
+    const result = [];
+    while (array.length > 0) {
+      const groupByNumber = array.length >= n ? n : array.length;
+      result.push(array.splice(0, groupByNumber));
+    }
+    return result;
+  };
 };
+const logError = console.error;
 
-class CameraRollView extends React.Component<Props, State> {
+class CameraRollView extends React.Component {
   static defaultProps = {
     groupTypes: 'All',
     batchSize: 5,
     imagesPerRow: 1,
     assetType: 'Photos',
-    renderImage: function(asset: PhotoIdentifier) {
+    renderImage: function(asset) {
       const imageSize = 150;
       const imageStyle = [styles.image, {width: imageSize, height: imageSize}];
       return <Image source={asset.node.image} style={imageStyle} />;
@@ -118,13 +68,13 @@ class CameraRollView extends React.Component<Props, State> {
     this.fetch();
   }
 
-  UNSAFE_componentWillReceiveProps(nextProps: Props) {
+  UNSAFE_componentWillReceiveProps(nextProps) {
     if (this.props.groupTypes !== nextProps.groupTypes) {
       this.fetch(true);
     }
   }
 
-  async _fetch(clear?: boolean) {
+  async _fetch(clear) {
     if (clear) {
       this.setState(this.getInitialState(), this.fetch);
       return;
@@ -144,7 +94,7 @@ class CameraRollView extends React.Component<Props, State> {
       }
     }
 
-    const fetchParams: GetPhotosParams = {
+    const fetchParams = {
       first: this.props.batchSize,
       groupTypes: this.props.groupTypes,
       assetType: this.props.assetType,
@@ -170,7 +120,7 @@ class CameraRollView extends React.Component<Props, State> {
    * Fetches more images from the camera roll. If clear is set to true, it will
    * set the component to its initial state and re-fetch the images.
    */
-  fetch = (clear?: boolean) => {
+  fetch = clear => {
     if (!this.state.loadingMore) {
       this.setState({loadingMore: true}, () => {
         this._fetch(clear);
@@ -179,6 +129,7 @@ class CameraRollView extends React.Component<Props, State> {
   };
 
   render() {
+    console.log({data: this.state.data});
     return (
       <FlatList
         keyExtractor={(_, idx) => String(idx)}
@@ -200,17 +151,18 @@ class CameraRollView extends React.Component<Props, State> {
     return null;
   };
 
-  _renderItem = (row: Row) => {
+  _renderItem = ({item}) => {
+    console.log({item});
     return (
       <View style={styles.row}>
-        {row.item.map(image => (image ? this.props.renderImage(image) : null))}
+        {item.map(image => (image ? this.props.renderImage(image) : null))}
       </View>
     );
   };
 
-  _appendAssets(data: PhotoIdentifiersPage) {
+  _appendAssets(data) {
     const assets = data.edges;
-    const newState: $Shape<State> = {loadingMore: false};
+    const newState = {loadingMore: false};
 
     if (!data.page_info.has_next_page) {
       newState.noMore = true;
@@ -234,10 +186,7 @@ class CameraRollView extends React.Component<Props, State> {
       }
 
       newState.assets = this.state.assets.concat(uniqAssets);
-      newState.data = groupByEveryN<PhotoIdentifier>(
-        newState.assets,
-        this.props.imagesPerRow,
-      );
+      newState.data = groupByEveryN(this.props.imagesPerRow)(newState.assets);
     }
 
     this.setState(newState);
