@@ -1,1 +1,4 @@
-module.exports = { extends: "@react-native-community" };
+module.exports = {
+  root: true,
+  extends: '@react-native-community',
+};
