@@ -8,6 +8,7 @@
 package com.reactnativecommunity.cameraroll;
 
 import android.content.ContentResolver;
+import android.content.ContentUris;
 import android.content.Context;
 import android.content.res.AssetFileDescriptor;
 import android.database.Cursor;
@@ -21,6 +22,7 @@ import android.os.Environment;
 import android.provider.MediaStore;
 import android.provider.MediaStore.Images;
 import android.text.TextUtils;
+import android.media.ExifInterface;
 
 import com.facebook.common.logging.FLog;
 import com.facebook.react.bridge.GuardedAsyncTask;
@@ -42,14 +44,19 @@ import com.facebook.react.module.annotations.ReactModule;
 
 import java.io.File;
 import java.io.FileInputStream;
+import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.nio.channels.FileChannel;
 import java.util.ArrayList;
+import java.util.HashSet;
 import java.util.List;
 import java.text.SimpleDateFormat;
 import java.text.ParseException;
 import java.util.Date;
+import java.util.Map;
+import java.util.HashMap;
+import java.util.Set;
 
 import javax.annotation.Nullable;
 
@@ -65,11 +72,18 @@ public class CameraRollModule extends ReactContextBaseJavaModule {
   private static final String ERROR_UNABLE_TO_LOAD = "E_UNABLE_TO_LOAD";
   private static final String ERROR_UNABLE_TO_LOAD_PERMISSION = "E_UNABLE_TO_LOAD_PERMISSION";
   private static final String ERROR_UNABLE_TO_SAVE = "E_UNABLE_TO_SAVE";
+  private static final String ERROR_UNABLE_TO_DELETE = "E_UNABLE_TO_DELETE";
   private static final String ERROR_UNABLE_TO_FILTER = "E_UNABLE_TO_FILTER";
 
   private static final String ASSET_TYPE_PHOTOS = "Photos";
   private static final String ASSET_TYPE_VIDEOS = "Videos";
   private static final String ASSET_TYPE_ALL = "All";
+
+  private static final String INCLUDE_FILENAME = "filename";
+  private static final String INCLUDE_FILE_SIZE = "fileSize";
+  private static final String INCLUDE_LOCATION = "location";
+  private static final String INCLUDE_IMAGE_SIZE = "imageSize";
+  private static final String INCLUDE_PLAYABLE_DURATION = "playableDuration";
 
   private static final String[] PROJECTION = {
     Images.Media._ID,
@@ -78,6 +92,7 @@ public class CameraRollModule extends ReactContextBaseJavaModule {
     Images.Media.DATE_TAKEN,
     MediaStore.MediaColumns.WIDTH,
     MediaStore.MediaColumns.HEIGHT,
+    MediaStore.MediaColumns.SIZE,
     MediaStore.MediaColumns.DATA
   };
 
@@ -127,10 +142,22 @@ public class CameraRollModule extends ReactContextBaseJavaModule {
       File source = new File(mUri.getPath());
       FileChannel input = null, output = null;
       try {
-        File environment = Environment.getExternalStoragePublicDirectory(
-          Environment.DIRECTORY_DCIM);
+        boolean isAlbumPresent = !"".equals(mOptions.getString("album"));
+        
+        final File environment;
+        // Media is not saved into an album when using Environment.DIRECTORY_DCIM.
+        if (isAlbumPresent) {
+          if ("video".equals(mOptions.getString("type"))) {
+            environment = Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_MOVIES);
+          } else {
+            environment = Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_PICTURES);
+          }
+        } else {
+          environment = Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_DCIM);
+        }
+
         File exportDir;
-        if (!"".equals(mOptions.getString("album"))) {
+        if (isAlbumPresent) {
           exportDir = new File(environment, mOptions.getString("album"));
           if (!exportDir.exists() && !exportDir.mkdirs()) {
             mPromise.reject(ERROR_UNABLE_TO_LOAD, "Album Directory not created. Did you request WRITE_EXTERNAL_STORAGE?");
@@ -228,9 +255,12 @@ public class CameraRollModule extends ReactContextBaseJavaModule {
     String after = params.hasKey("after") ? params.getString("after") : null;
     String groupName = params.hasKey("groupName") ? params.getString("groupName") : null;
     String assetType = params.hasKey("assetType") ? params.getString("assetType") : ASSET_TYPE_PHOTOS;
+    long fromTime = params.hasKey("fromTime") ? (long) params.getDouble("fromTime") : 0;
+    long toTime = params.hasKey("toTime") ? (long) params.getDouble("toTime") : 0;
     ReadableArray mimeTypes = params.hasKey("mimeTypes")
         ? params.getArray("mimeTypes")
         : null;
+    ReadableArray include = params.hasKey("include") ? params.getArray("include") : null;
 
     new GetMediaTask(
           getReactApplicationContext(),
@@ -239,6 +269,9 @@ public class CameraRollModule extends ReactContextBaseJavaModule {
           groupName,
           mimeTypes,
           assetType,
+          fromTime,
+          toTime,
+          include,
           promise)
           .executeOnExecutor(AsyncTask.THREAD_POOL_EXECUTOR);
   }
@@ -251,6 +284,9 @@ public class CameraRollModule extends ReactContextBaseJavaModule {
     private final @Nullable ReadableArray mMimeTypes;
     private final Promise mPromise;
     private final String mAssetType;
+    private final long mFromTime;
+    private final long mToTime;
+    private final Set<String> mInclude;
 
     private GetMediaTask(
         ReactContext context,
@@ -259,6 +295,9 @@ public class CameraRollModule extends ReactContextBaseJavaModule {
         @Nullable String groupName,
         @Nullable ReadableArray mimeTypes,
         String assetType,
+        long fromTime,
+        long toTime,
+        @Nullable ReadableArray include,
         Promise promise) {
       super(context);
       mContext = context;
@@ -268,16 +307,32 @@ public class CameraRollModule extends ReactContextBaseJavaModule {
       mMimeTypes = mimeTypes;
       mPromise = promise;
       mAssetType = assetType;
+      mFromTime = fromTime;
+      mToTime = toTime;
+      mInclude = createSetFromIncludeArray(include);
+    }
+
+    private static Set<String> createSetFromIncludeArray(@Nullable ReadableArray includeArray) {
+      Set<String> includeSet = new HashSet<>();
+
+      if (includeArray == null) {
+        return includeSet;
+      }
+
+      for (int i = 0; i < includeArray.size(); i++) {
+        @Nullable String includeItem = includeArray.getString(i);
+        if (includeItem != null) {
+          includeSet.add(includeItem);
+        }
+      }
+
+      return includeSet;
     }
 
     @Override
     protected void doInBackgroundGuarded(Void... params) {
       StringBuilder selection = new StringBuilder("1");
       List<String> selectionArgs = new ArrayList<>();
-      if (!TextUtils.isEmpty(mAfter)) {
-        selection.append(" AND " + SELECTION_DATE_TAKEN);
-        selectionArgs.add(mAfter);
-      }
       if (!TextUtils.isEmpty(mGroupName)) {
         selection.append(" AND " + SELECTION_BUCKET);
         selectionArgs.add(mGroupName);
@@ -311,25 +366,39 @@ public class CameraRollModule extends ReactContextBaseJavaModule {
         }
         selection.replace(selection.length() - 1, selection.length(), ")");
       }
+
+      if (mFromTime > 0) {
+        selection.append(" AND " + Images.Media.DATE_TAKEN + " > ?");
+        selectionArgs.add(mFromTime + "");
+      }
+      if (mToTime > 0) {
+        selection.append(" AND " + Images.Media.DATE_TAKEN + " <= ?");
+        selectionArgs.add(mToTime + "");
+      }
+
       WritableMap response = new WritableNativeMap();
       ContentResolver resolver = mContext.getContentResolver();
-      // using LIMIT in the sortOrder is not explicitly supported by the SDK (which does not support
-      // setting a limit at all), but it works because this specific ContentProvider is backed by
-      // an SQLite DB and forwards parameters to it without doing any parsing / validation.
+
       try {
+        // set LIMIT to first + 1 so that we know how to populate page_info
+        String limit = "limit=" + (mFirst + 1);
+
+        if (!TextUtils.isEmpty(mAfter)) {
+          limit = "limit=" + mAfter + "," + (mFirst + 1);
+        }
+
         Cursor media = resolver.query(
-            MediaStore.Files.getContentUri("external"),
+            MediaStore.Files.getContentUri("external").buildUpon().encodedQuery(limit).build(),
             PROJECTION,
             selection.toString(),
             selectionArgs.toArray(new String[selectionArgs.size()]),
-            Images.Media.DATE_TAKEN + " DESC, " + Images.Media.DATE_MODIFIED + " DESC LIMIT " +
-                (mFirst + 1)); // set LIMIT to first + 1 so that we know how to populate page_info
+            Images.Media.DATE_ADDED + " DESC, " + Images.Media.DATE_MODIFIED + " DESC");
         if (media == null) {
           mPromise.reject(ERROR_UNABLE_TO_LOAD, "Could not get media");
         } else {
           try {
-            putEdges(resolver, media, response, mFirst);
-            putPageInfo(media, response, mFirst);
+            putEdges(resolver, media, response, mFirst, mInclude);
+            putPageInfo(media, response, mFirst, !TextUtils.isEmpty(mAfter) ? Integer.parseInt(mAfter) : 0);
           } finally {
             media.close();
             mPromise.resolve(response);
@@ -344,14 +413,83 @@ public class CameraRollModule extends ReactContextBaseJavaModule {
     }
   }
 
-  private static void putPageInfo(Cursor media, WritableMap response, int limit) {
+  @ReactMethod
+  public void getAlbums(final ReadableMap params, final Promise promise) {
+    String assetType = params.hasKey("assetType") ? params.getString("assetType") : ASSET_TYPE_ALL;
+    StringBuilder selection = new StringBuilder("1");
+    List<String> selectionArgs = new ArrayList<>();
+    if (assetType.equals(ASSET_TYPE_PHOTOS)) {
+      selection.append(" AND " + MediaStore.Files.FileColumns.MEDIA_TYPE + " = "
+              + MediaStore.Files.FileColumns.MEDIA_TYPE_IMAGE);
+    } else if (assetType.equals(ASSET_TYPE_VIDEOS)) {
+      selection.append(" AND " + MediaStore.Files.FileColumns.MEDIA_TYPE + " = "
+              + MediaStore.Files.FileColumns.MEDIA_TYPE_VIDEO);
+    } else if (assetType.equals(ASSET_TYPE_ALL)) {
+      selection.append(" AND " + MediaStore.Files.FileColumns.MEDIA_TYPE + " IN ("
+              + MediaStore.Files.FileColumns.MEDIA_TYPE_VIDEO + ","
+              + MediaStore.Files.FileColumns.MEDIA_TYPE_IMAGE + ")");
+    } else {
+      promise.reject(
+              ERROR_UNABLE_TO_FILTER,
+              "Invalid filter option: '" + assetType + "'. Expected one of '"
+                      + ASSET_TYPE_PHOTOS + "', '" + ASSET_TYPE_VIDEOS + "' or '" + ASSET_TYPE_ALL + "'."
+      );
+      return;
+    }
+
+    final String[] projection = {MediaStore.Images.ImageColumns.BUCKET_DISPLAY_NAME};
+
+    try {
+      Cursor media = getReactApplicationContext().getContentResolver().query(
+              MediaStore.Files.getContentUri("external"),
+              projection,
+              selection.toString(),
+              selectionArgs.toArray(new String[selectionArgs.size()]),
+              null);
+      if (media == null) {
+        promise.reject(ERROR_UNABLE_TO_LOAD, "Could not get media");
+      } else {
+        WritableArray response = new WritableNativeArray();
+        try {
+          if (media.moveToFirst()) {
+            Map<String, Integer> albums = new HashMap<>();
+            do {
+              String albumName = media.getString(media.getColumnIndex(MediaStore.Images.ImageColumns.BUCKET_DISPLAY_NAME));
+              if (albumName != null) {
+                Integer albumCount = albums.get(albumName);
+                if (albumCount == null) {
+                  albums.put(albumName, 1);
+                } else {
+                  albums.put(albumName, albumCount + 1);
+                }
+              }
+            } while (media.moveToNext());
+
+            for (Map.Entry<String, Integer> albumEntry : albums.entrySet()) {
+              WritableMap album = new WritableNativeMap();
+              album.putString("title", albumEntry.getKey());
+              album.putInt("count", albumEntry.getValue());
+              response.pushMap(album);
+            }
+          }
+        } finally {
+          media.close();
+          promise.resolve(response);
+        }
+      }
+    } catch (Exception e) {
+      promise.reject(ERROR_UNABLE_TO_LOAD, "Could not get media", e);
+    }
+  }
+
+  private static void putPageInfo(Cursor media, WritableMap response, int limit, int offset) {
     WritableMap pageInfo = new WritableNativeMap();
     pageInfo.putBoolean("has_next_page", limit < media.getCount());
     if (limit < media.getCount()) {
-      media.moveToPosition(limit - 1);
       pageInfo.putString(
-          "end_cursor",
-          media.getString(media.getColumnIndex(Images.Media.DATE_TAKEN)));
+        "end_cursor",
+        Integer.toString(offset + limit)
+      );
     }
     response.putMap("page_info", pageInfo);
   }
@@ -360,23 +498,32 @@ public class CameraRollModule extends ReactContextBaseJavaModule {
       ContentResolver resolver,
       Cursor media,
       WritableMap response,
-      int limit) {
+      int limit,
+      Set<String> include) {
     WritableArray edges = new WritableNativeArray();
     media.moveToFirst();
-    int idIndex = media.getColumnIndex(Images.Media._ID);
     int mimeTypeIndex = media.getColumnIndex(Images.Media.MIME_TYPE);
     int groupNameIndex = media.getColumnIndex(Images.Media.BUCKET_DISPLAY_NAME);
     int dateTakenIndex = media.getColumnIndex(Images.Media.DATE_TAKEN);
     int widthIndex = media.getColumnIndex(MediaStore.MediaColumns.WIDTH);
     int heightIndex = media.getColumnIndex(MediaStore.MediaColumns.HEIGHT);
+    int sizeIndex = media.getColumnIndex(MediaStore.MediaColumns.SIZE);
     int dataIndex = media.getColumnIndex(MediaStore.MediaColumns.DATA);
+
+    boolean includeLocation = include.contains(INCLUDE_LOCATION);
+    boolean includeFilename = include.contains(INCLUDE_FILENAME);
+    boolean includeFileSize = include.contains(INCLUDE_FILE_SIZE);
+    boolean includeImageSize = include.contains(INCLUDE_IMAGE_SIZE);
+    boolean includePlayableDuration = include.contains(INCLUDE_PLAYABLE_DURATION);
 
     for (int i = 0; i < limit && !media.isAfterLast(); i++) {
       WritableMap edge = new WritableNativeMap();
       WritableMap node = new WritableNativeMap();
       ExifInterface exif = getExifInterface(media, dataIndex);
       boolean imageInfoSuccess = exif != null &&
-          putImageInfo(resolver, media, node, idIndex, widthIndex, heightIndex, dataIndex, mimeTypeIndex, exif);
+          putImageInfo(resolver, media, node, widthIndex, heightIndex, sizeIndex, dataIndex,
+              mimeTypeIndex, includeFilename, includeFileSize, includeImageSize,
+              includePlayableDuration, exif);
       if (imageInfoSuccess) {
         putBasicNodeInfo(media, node, mimeTypeIndex, groupNameIndex, dateTakenIndex);
         putLocationInfo(media, node, dataIndex, exif);
@@ -416,83 +563,180 @@ public class CameraRollModule extends ReactContextBaseJavaModule {
     node.putDouble("timestamp", media.getLong(dateTakenIndex) / 1000d);
   }
 
+  /**
+   * @return Whether we successfully fetched all the information about the image that we were asked
+   * to include
+   */
   private static boolean putImageInfo(
       ContentResolver resolver,
       Cursor media,
       WritableMap node,
-      int idIndex,
       int widthIndex,
       int heightIndex,
+      int sizeIndex,
       int dataIndex,
       int mimeTypeIndex,
       ExifInterface exif) {
+      boolean includeFilename,
+      boolean includeFileSize,
+      boolean includeImageSize,
+      boolean includePlayableDuration) {
     WritableMap image = new WritableNativeMap();
     Uri photoUri = Uri.parse("file://" + media.getString(dataIndex));
-    File file = new File(media.getString(dataIndex));
-    String strFileName = file.getName();
     image.putString("uri", photoUri.toString());
-    image.putString("filename", strFileName);
-    float width = media.getInt(widthIndex);
-    float height = media.getInt(heightIndex);
-
     String mimeType = media.getString(mimeTypeIndex);
 
-    if (mimeType != null
-        && mimeType.startsWith("video")) {
-      try {
-        AssetFileDescriptor photoDescriptor = resolver.openAssetFileDescriptor(photoUri, "r");
-        MediaMetadataRetriever retriever = new MediaMetadataRetriever();
-        retriever.setDataSource(photoDescriptor.getFileDescriptor());
+    boolean isVideo = mimeType != null && mimeType.startsWith("video");
+    boolean putImageSizeSuccess = putImageSize(resolver, media, image, widthIndex, heightIndex,
+        photoUri, isVideo, includeImageSize);
+    boolean putPlayableDurationSuccess = putPlayableDuration(resolver, image, photoUri, isVideo,
+        includePlayableDuration);
 
-        try {
-          if (width <= 0 || height <= 0) {
+    if (includeFilename) {
+      File file = new File(media.getString(dataIndex));
+      String strFileName = file.getName();
+      image.putString("filename", strFileName);
+    } else {
+      image.putNull("filename");
+    }
+
+    if (includeFileSize) {
+      image.putDouble("fileSize", media.getLong(sizeIndex));
+    } else {
+      image.putNull("fileSize");
+    }
+
+    node.putMap("image", image);
+    return putImageSizeSuccess && putPlayableDurationSuccess;
+  }
+
+  /**
+   * @return Whether we succeeded in fetching and putting the playableDuration
+   */
+  private static boolean putPlayableDuration(
+      ContentResolver resolver,
+      WritableMap image,
+      Uri photoUri,
+      boolean isVideo,
+      boolean includePlayableDuration) {
+    image.putNull("playableDuration");
+
+    if (!includePlayableDuration || !isVideo) {
+      return true;
+    }
+
+    boolean success = true;
+    @Nullable Integer playableDuration = null;
+    @Nullable AssetFileDescriptor photoDescriptor = null;
+    try {
+      photoDescriptor = resolver.openAssetFileDescriptor(photoUri, "r");
+    } catch (FileNotFoundException e) {
+      success = false;
+      FLog.e(ReactConstants.TAG, "Could not open asset file " + photoUri.toString(), e);
+    }
+
+    if (photoDescriptor != null) {
+      MediaMetadataRetriever retriever = new MediaMetadataRetriever();
+      retriever.setDataSource(photoDescriptor.getFileDescriptor());
+      try {
+        int timeInMillisec =
+            Integer.parseInt(
+                retriever.extractMetadata(MediaMetadataRetriever.METADATA_KEY_DURATION));
+        playableDuration = timeInMillisec / 1000;
+      } catch (NumberFormatException e) {
+        success = false;
+        FLog.e(
+            ReactConstants.TAG,
+            "Number format exception occurred while trying to fetch video metadata for "
+                + photoUri.toString(),
+            e);
+      }
+      retriever.release();
+    }
+
+    if (photoDescriptor != null) {
+      try {
+        photoDescriptor.close();
+      } catch (IOException e) {
+        // Do nothing. We can't handle this, and this is usually a system problem
+      }
+    }
+
+    if (playableDuration != null) {
+      image.putInt("playableDuration", playableDuration);
+    }
+
+    return success;
+  }
+
+  private static boolean putImageSize(
+      ContentResolver resolver,
+      Cursor media,
+      WritableMap image,
+      int widthIndex,
+      int heightIndex,
+      Uri photoUri,
+      boolean isVideo,
+      boolean includeImageSize) {
+    image.putNull("width");
+    image.putNull("height");
+
+    if (!includeImageSize) {
+      return true;
+    }
+
+    boolean success = true;
+    int width = media.getInt(widthIndex);
+    int height = media.getInt(heightIndex);
+
+    if (width <= 0 || height <= 0) {
+      @Nullable AssetFileDescriptor photoDescriptor = null;
+      try {
+        photoDescriptor = resolver.openAssetFileDescriptor(photoUri, "r");
+      } catch (FileNotFoundException e) {
+        success = false;
+        FLog.e(ReactConstants.TAG, "Could not open asset file " + photoUri.toString(), e);
+      }
+
+      if (photoDescriptor != null) {
+        if (isVideo) {
+          MediaMetadataRetriever retriever = new MediaMetadataRetriever();
+          retriever.setDataSource(photoDescriptor.getFileDescriptor());
+          try {
             width =
                 Integer.parseInt(
                     retriever.extractMetadata(MediaMetadataRetriever.METADATA_KEY_VIDEO_WIDTH));
             height =
                 Integer.parseInt(
                     retriever.extractMetadata(MediaMetadataRetriever.METADATA_KEY_VIDEO_HEIGHT));
+          } catch (NumberFormatException e) {
+            success = false;
+            FLog.e(
+                ReactConstants.TAG,
+                "Number format exception occurred while trying to fetch video metadata for "
+                    + photoUri.toString(),
+                e);
           }
-          int timeInMillisec =
-              Integer.parseInt(
-                  retriever.extractMetadata(MediaMetadataRetriever.METADATA_KEY_DURATION));
-          int playableDuration = timeInMillisec / 1000;
-          image.putInt("playableDuration", playableDuration);
-        } catch (NumberFormatException e) {
-          FLog.e(
-              ReactConstants.TAG,
-              "Number format exception occurred while trying to fetch video metadata for "
-                  + photoUri.toString(),
-              e);
-          return false;
-        } finally {
           retriever.release();
-          photoDescriptor.close();
+        } else {
+          BitmapFactory.Options options = new BitmapFactory.Options();
+          // Set inJustDecodeBounds to true so we don't actually load the Bitmap, but only get its
+          // dimensions instead.
+          options.inJustDecodeBounds = true;
+          BitmapFactory.decodeFileDescriptor(photoDescriptor.getFileDescriptor(), null, options);
+          width = options.outWidth;
+          height = options.outHeight;
         }
-      } catch (Exception e) {
-        FLog.e(ReactConstants.TAG, "Could not get video metadata for " + photoUri.toString(), e);
-        return false;
       }
-    }
 
-    if (width <= 0 || height <= 0) {
       try {
-        AssetFileDescriptor photoDescriptor = resolver.openAssetFileDescriptor(photoUri, "r");
-        BitmapFactory.Options options = new BitmapFactory.Options();
-        // Set inJustDecodeBounds to true so we don't actually load the Bitmap, but only get its
-        // dimensions instead.
-        options.inJustDecodeBounds = true;
-        BitmapFactory.decodeFileDescriptor(photoDescriptor.getFileDescriptor(), null, options);
-        width = options.outWidth;
-        height = options.outHeight;
         photoDescriptor.close();
       } catch (IOException e) {
-        FLog.e(ReactConstants.TAG, "Could not get width/height for " + photoUri.toString(), e);
-        return false;
+        // Do nothing. We can't handle this, and this is usually a system problem
       }
     }
-    image.putDouble("width", width);
-    image.putDouble("height", height);
+    image.putInt("width", width);
+    image.putInt("height", height);
     node.putMap("image", image);
     try {
       String exifTimestampString = exif.getAttribute("DateTime");
@@ -503,9 +747,10 @@ public class CameraRollModule extends ReactContextBaseJavaModule {
       }
     } catch (ParseException e) {
       FLog.e(ReactConstants.TAG, "Could not parse exifTimestamp for " + photoUri.toString(), e);
-      return false;
+      success = false;
     }
-    return true;
+
+    return success;
   }
 
   private static void putLocationInfo(
@@ -526,5 +771,80 @@ public class CameraRollModule extends ReactContextBaseJavaModule {
         location.putDouble("latitude", latitude);
         node.putMap("location", location);
       }
+  }
+
+  /**
+   * Delete a set of images.
+   *
+   * @param uris array of file:// URIs of the images to delete
+   * @param promise to be resolved
+   */
+  @ReactMethod
+  public void deletePhotos(ReadableArray uris, Promise promise) {
+    if (uris.size() == 0) {
+      promise.reject(ERROR_UNABLE_TO_DELETE, "Need at least one URI to delete");
+    } else {
+      new DeletePhotos(getReactApplicationContext(), uris, promise)
+          .executeOnExecutor(AsyncTask.THREAD_POOL_EXECUTOR);
+    }
+  }
+
+  private static class DeletePhotos extends GuardedAsyncTask<Void, Void> {
+
+    private final Context mContext;
+    private final ReadableArray mUris;
+    private final Promise mPromise;
+
+    public DeletePhotos(ReactContext context, ReadableArray uris, Promise promise) {
+      super(context);
+      mContext = context;
+      mUris = uris;
+      mPromise = promise;
+    }
+
+    @Override
+    protected void doInBackgroundGuarded(Void... params) {
+      ContentResolver resolver = mContext.getContentResolver();
+
+      // Set up the projection (we only need the ID)
+      String[] projection = { MediaStore.Images.Media._ID };
+
+      // Match on the file path
+      String innerWhere = "?";
+      for (int i = 1; i < mUris.size(); i++) {
+        innerWhere += ", ?";
+      }
+
+      String selection = MediaStore.Images.Media.DATA + " IN (" + innerWhere + ")";
+      // Query for the ID of the media matching the file path
+      Uri queryUri = MediaStore.Images.Media.EXTERNAL_CONTENT_URI;
+
+      String[] selectionArgs = new String[mUris.size()];
+      for (int i = 0; i < mUris.size(); i++) {
+        Uri uri = Uri.parse(mUris.getString(i));
+        selectionArgs[i] = uri.getPath();
+      }
+
+      Cursor cursor = resolver.query(queryUri, projection, selection, selectionArgs, null);
+      int deletedCount = 0;
+
+      while (cursor.moveToNext()) {
+        long id = cursor.getLong(cursor.getColumnIndexOrThrow(MediaStore.Images.Media._ID));
+        Uri deleteUri = ContentUris.withAppendedId(MediaStore.Images.Media.EXTERNAL_CONTENT_URI, id);
+
+        if (resolver.delete(deleteUri, null, null) == 1) {
+          deletedCount++;
+        }
+      }
+
+      cursor.close();
+
+      if (deletedCount == mUris.size()) {
+        mPromise.resolve(true);
+      } else {
+        mPromise.reject(ERROR_UNABLE_TO_DELETE,
+            "Could not delete all media, only deleted " + deletedCount + " photos.");
+      }
+    }
   }
 }
